refactor: Redo whole gallery section
This commit will refactor the whole gallery section, since the
previous approach did not work, was not responsive and since
the gallery contains pictures with different resolutions, it did not
work. The current solution looks great. The following changes are made:

- Change html gallery section to a list, each pic is a list item
- Add generic flexbox container and boxes, which wrap
- Set min height to each pic and flex basis width
- Set a min height to each pic to look good in various sizes
- Give each li / pic the setting to use it's whole "box"

This commit will also remove some comments out of the code of the
nav-list section