feat: Add content to confirmation page
The confirmation page shares nav, header, footer with the tours
page and therefore the code has just been copied without single
individual commits.

However, a few styling tweaks are done in order to have the
hero image and the text inside correctly displayed in all
sizes.
The display text got a partially transparent box, position is
lower than on the other hero pages and also the font gap got
changed. For this dedicated case, the font of the paragraph got
changed to match headings.

- Copy header, nav, footer from tours page
- Add hero image and put opacity filter on it
- Change style to place text better
- Change style of the text
- Add transparent box around title text