fix: Add missing div character & change YT link
This commit will add the fix for a missing div closing property, which
was missing in index from line 43 facts container.

It will also replace the YT link, which I did modify via the
https://www.classynemesis.com/projects/ytembed/. I did not like
the fact to be dependant on this page, sicne the YT link was
not the original one.

- Add closing div property for line 43 opening to index.html
- Replace YT link to original one & adapt related CSS rule for position