docs: Update readme and testing styling
This commit inserts a nice effect to the Wireframe images on the
readme file, via the details tag to open / close the iamge behind.

It also removes the table of content on readme and testing, since
I've figured that on small screen sizes it is not of big help and
the already embedded github ToC via hamburger menu, works better.

- Add details tag / feature to wireframe images
- Remove table of content from readme and testing file