docs: Add format corrections to readme / testing
During review of the readme and testing file, I've figured that
I did inconsisten file naming, including capital letters in image
file names. This commit will fix that error to all lowercase and
update the readme and testing file.

This commit will also add some styling corrections in the docs and
add the desktop responsiveness images to the testing file.

- Correct image naming to all lowercase
- Update readme & testing with corrected image names
- Style bullet list in readme
- Add desktop responsive testing section to testing file